add queue dependency, use queue to provide sequential execution of requests
